requirements.txt
requests
bs4
textblob
pandas
pylast
python-dotenv